requirements.txt
gradio==4.3.0
matplotlib==3.4.3
numpy==1.23.4
numpy==1.22.3
opencv_contrib_python==4.6.0.66
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.0.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
